requirements.txt
streamlit
langchain
langchain_community
openai
bs4
PyPDF2
requests
python-dotenv
